refactor: ExistenceCheckAdvice 이슈 해결
